Add main and rename lobe to a header
Added a main.cpp that compiles and prints a simple message. Renamed lobe.cpp to
lobe.hpp. Corrected syntax to allow compilation. This program prints and exits.
The current chosen type for many of the private variables in the lobe class
might require change in the future. Decoupling the variable type for these
privates can be done by creating a concentration type/class. This type can hide
the elementary types underneath

Signed-off-by: Ahmed Mansour <ahmed.om@gmail.com>