Restructure main to use Lobe object
Add the lobe object in main program to manage large groups of cells. Soft
threads will be used to manage actions in each cell. This implementation detail
will not be visible at the lobe level. Lobes will be used to manage cell groups
though their lifecycle from creation to death. Meanigful stimuly and testing
will be done at the lobe level. This commit breaks the compilation as the Lobe
class is still to be implemented

Signed-off-by: Ahmed Mansour <ahmed.om@gmail.com>